+++ etherman/etherman.go
@@ -2,7 +2,6 @@ package etherman
 
 import (
 	"context"
-	"crypto/ecdsa"
 	"encoding/hex"
 	"encoding/json"
 	"errors"
@@ -530,14 +529,14 @@ func (etherMan *Client) WaitTxToBeMined(ctx context.Context, tx *types.Transacti
 }
 
 // EstimateGasSequenceBatches estimates gas for sending batches
-func (etherMan *Client) EstimateGasSequenceBatches(sender common.Address, sequences []ethmanTypes.Sequence, l2Coinbase common.Address, committeeSignaturesAndAddrs []byte) (*types.Transaction, error) {
-	opts, err := etherMan.generateMockAuth(sender)
+func (etherMan *Client) EstimateGasSequenceBatches(sender common.Address, sequences []ethmanTypes.Sequence, l2Coinbase common.Address) (*types.Transaction, error) {
+	opts, err := etherMan.getAuthByAddress(sender)
 	if err == ErrNotFound {
 		return nil, ErrPrivateKeyNotFound
 	}
 	opts.NoSend = true
 
-	tx, err := etherMan.sequenceBatches(opts, sequences, l2Coinbase, committeeSignaturesAndAddrs)
+	tx, err := etherMan.sequenceBatches(opts, sequences, l2Coinbase)
 	if err != nil {
 		return nil, err
 	}
@@ -546,8 +545,8 @@ func (etherMan *Client) EstimateGasSequenceBatches(sender common.Address, sequen
 }
 
 // BuildSequenceBatchesTxData builds a []bytes to be sent to the PoE SC method SequenceBatches.
-func (etherMan *Client) BuildSequenceBatchesTxData(sender common.Address, sequences []ethmanTypes.Sequence, l2Coinbase common.Address, committeeSignaturesAndAddrs []byte) (to *common.Address, data []byte, err error) {
-	opts, err := etherMan.generateRandomAuth()
+func (etherMan *Client) BuildSequenceBatchesTxData(sender common.Address, sequences []ethmanTypes.Sequence, l2Coinbase common.Address) (to *common.Address, data []byte, err error) {
+	opts, err := etherMan.getAuthByAddress(sender)
 	if err == ErrNotFound {
 		return nil, nil, fmt.Errorf("failed to build sequence batches, err: %w", ErrPrivateKeyNotFound)
 	}
@@ -557,7 +556,7 @@ func (etherMan *Client) BuildSequenceBatchesTxData(sender common.Address, sequen
 	opts.GasLimit = uint64(1)
 	opts.GasPrice = big.NewInt(1)
 
-	tx, err := etherMan.sequenceBatches(opts, sequences, l2Coinbase, committeeSignaturesAndAddrs)
+	tx, err := etherMan.sequenceBatches(opts, sequences, l2Coinbase)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -565,49 +564,24 @@ func (etherMan *Client) BuildSequenceBatchesTxData(sender common.Address, sequen
 	return tx.To(), tx.Data(), nil
 }
 
-func (etherMan *Client) sequenceBatches(opts bind.TransactOpts, sequences []ethmanTypes.Sequence, l2Coinbase common.Address, committeeSignaturesAndAddrs []byte) (*types.Transaction, error) {
+func (etherMan *Client) sequenceBatches(opts bind.TransactOpts, sequences []ethmanTypes.Sequence, l2Coinbase common.Address) (*types.Transaction, error) {
 	var batches []polygonzkevm.PolygonZkEVMBatchData
-
-	var tx *types.Transaction
-	var err error
-	if len(committeeSignaturesAndAddrs) > 0 {
-		for _, seq := range sequences {
-			batch := polygonzkevm.PolygonZkEVMBatchData{
-				TransactionsHash:   crypto.Keccak256Hash(seq.BatchL2Data),
-				GlobalExitRoot:     seq.GlobalExitRoot,
-				Timestamp:          uint64(seq.Timestamp),
-				MinForcedTimestamp: uint64(seq.ForcedBatchTimestamp),
-			}
-
-			batches = append(batches, batch)
+	for _, seq := range sequences {
+		batch := polygonzkevm.PolygonZkEVMBatchData{
+			Transactions:       seq.BatchL2Data,
+			GlobalExitRoot:     seq.GlobalExitRoot,
+			Timestamp:          uint64(seq.Timestamp),
+			MinForcedTimestamp: uint64(seq.ForcedBatchTimestamp),
 		}
 
-		log.Infof("Sequence batches with validium.")
-		tx, err = etherMan.ZkEVM.SequenceBatches(&opts, batches, l2Coinbase, committeeSignaturesAndAddrs)
-	} else {
-		for _, seq := range sequences {
-			batch := polygonzkevm.PolygonZkEVMBatchData{
-				Transactions:       seq.BatchL2Data,
-				GlobalExitRoot:     seq.GlobalExitRoot,
-				Timestamp:          uint64(seq.Timestamp),
-				MinForcedTimestamp: uint64(seq.ForcedBatchTimestamp),
-			}
-
-			batches = append(batches, batch)
-		}
-
-		log.Infof("Sequence batches with rollup.")
-		tx, err = etherMan.ZkEVM.SequenceBatches(&opts, batches, l2Coinbase, nil)
+		batches = append(batches, batch)
 	}
 
+	tx, err := etherMan.ZkEVM.SequenceBatches(&opts, batches, l2Coinbase, nil) //TODO:scf ?????
 	if err != nil {
 		if parsedErr, ok := tryParseError(err); ok {
 			err = parsedErr
 		}
-		err = fmt.Errorf(
-			"error sequencing batches: %w, committeeSignaturesAndAddrs %s",
-			err, common.Bytes2Hex(committeeSignaturesAndAddrs),
-		)
 	}
 
 	return tx, err
@@ -1203,15 +1177,15 @@ func (etherMan *Client) AddOrReplaceAuth(auth bind.TransactOpts) error {
 }
 
 // LoadAuthFromKeyStore loads an authorization from a key store file
-func (etherMan *Client) LoadAuthFromKeyStore(path, password string) (*bind.TransactOpts, *ecdsa.PrivateKey, error) {
-	auth, pk, err := newAuthFromKeystore(path, password, etherMan.l1Cfg.L1ChainID)
+func (etherMan *Client) LoadAuthFromKeyStore(path, password string) (*bind.TransactOpts, error) {
+	auth, err := newAuthFromKeystore(path, password, etherMan.l1Cfg.L1ChainID)
 	if err != nil {
-		return nil, nil, err
+		return nil, err
 	}
 
 	log.Infof("loaded authorization for address: %v", auth.From.String())
 	etherMan.auth[auth.From] = auth
-	return &auth, pk, nil
+	return &auth, nil
 }
 
 // newKeyFromKeystore creates an instance of a keystore key from a keystore file
@@ -1232,20 +1206,20 @@ func newKeyFromKeystore(path, password string) (*keystore.Key, error) {
 }
 
 // newAuthFromKeystore an authorization instance from a keystore file
-func newAuthFromKeystore(path, password string, chainID uint64) (bind.TransactOpts, *ecdsa.PrivateKey, error) {
+func newAuthFromKeystore(path, password string, chainID uint64) (bind.TransactOpts, error) {
 	log.Infof("reading key from: %v", path)
 	key, err := newKeyFromKeystore(path, password)
 	if err != nil {
-		return bind.TransactOpts{}, nil, err
+		return bind.TransactOpts{}, err
 	}
 	if key == nil {
-		return bind.TransactOpts{}, nil, nil
+		return bind.TransactOpts{}, nil
 	}
 	auth, err := bind.NewKeyedTransactorWithChainID(key.PrivateKey, new(big.Int).SetUint64(chainID))
 	if err != nil {
-		return bind.TransactOpts{}, nil, err
+		return bind.TransactOpts{}, err
 	}
-	return *auth, key.PrivateKey, nil
+	return *auth, nil
 }
 
 // getAuthByAddress tries to get an authorization from the authorizations map
@@ -1271,35 +1245,5 @@ func (etherMan *Client) generateRandomAuth() (bind.TransactOpts, error) {
 		return bind.TransactOpts{}, errors.New("failed to generate a fake authorization to estimate L1 txs")
 	}
 
-	return *auth, nil
-}
-
-// generateMockAuth generates an authorization instance from a
-// randomly generated private key to be used to estimate gas for PoE
-// operations NOT restricted to the Trusted Sequencer
-func (etherMan *Client) generateMockAuth(sender common.Address) (bind.TransactOpts, error) {
-	privateKey, err := crypto.GenerateKey()
-	if err != nil {
-		return bind.TransactOpts{}, errors.New("failed to generate a private key to estimate L1 txs")
-	}
-	chainID := big.NewInt(0).SetUint64(etherMan.l1Cfg.L1ChainID)
-	auth, err := bind.NewKeyedTransactorWithChainID(privateKey, chainID)
-	if err != nil {
-		return bind.TransactOpts{}, errors.New("failed to generate a fake authorization to estimate L1 txs")
-	}
-
-	auth.From = sender
-	auth.Signer = func(address common.Address, tx *types.Transaction) (*types.Transaction, error) {
-		chainID := big.NewInt(0).SetUint64(etherMan.l1Cfg.L1ChainID)
-		signer := types.LatestSignerForChainID(chainID)
-		if err != nil {
-			return nil, err
-		}
-		signature, err := crypto.Sign(signer.Hash(tx).Bytes(), privateKey)
-		if err != nil {
-			return nil, err
-		}
-		return tx.WithSignature(signer, signature)
-	}
 	return *auth, nil
 }
