+++ sequencer/metrics/logstatisticsimpl.go
@@ -46,14 +46,6 @@ func (l *logStatisticsInstance) SetTag(tag logTag, value string) {
 	l.tags[tag] = value
 }
 
-func (l *logStatisticsInstance) GetTag(tag logTag) string {
-	return l.tags[tag]
-}
-
-func (l *logStatisticsInstance) GetStatistics(tag logTag) int64 {
-	return l.statistics[tag]
-}
-
 func (l *logStatisticsInstance) UpdateTimestamp(tag logTag, tm time.Time) {
 	l.timestamp[tag] = tm
 }
