+++ sequencesender/sequencesender_x1.go
@@ -1,0 +1,224 @@
+package sequencesender
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"github.com/0xPolygonHermez/zkevm-node/etherman/types"
+	"github.com/0xPolygonHermez/zkevm-node/ethtxmanager"
+	"github.com/0xPolygonHermez/zkevm-node/log"
+	"github.com/0xPolygonHermez/zkevm-node/sequencer/metrics"
+	"github.com/0xPolygonHermez/zkevm-node/state"
+	ethTypes "github.com/ethereum/go-ethereum/core/types"
+	"github.com/jackc/pgx/v4"
+	"time"
+)
+
+func (s *SequenceSender) tryToSendSequenceX1(ctx context.Context, ticker *time.Ticker) {
+	retry := false
+	// process monitored sequences before starting a next cycle
+	s.ethTxManager.ProcessPendingMonitoredTxs(ctx, ethTxManagerOwner, func(result ethtxmanager.MonitoredTxResult, dbTx pgx.Tx) {
+		if result.Status == ethtxmanager.MonitoredTxStatusFailed {
+			retry = true
+			mTxResultLogger := ethtxmanager.CreateMonitoredTxResultLogger(ethTxManagerOwner, result)
+			mTxResultLogger.Error("failed to send sequence, TODO: review this fatal and define what to do in this case")
+		}
+	}, nil)
+
+	if retry {
+		waitTick(ctx, ticker)
+		return
+	}
+
+	// Check if synchronizer is up to date
+	if !s.isSynced(ctx) {
+		log.Info("wait for synchronizer to sync last batch")
+		waitTick(ctx, ticker)
+		return
+	}
+
+	// Check if should send sequence to L1
+	log.Infof("getting sequences to send")
+	sequences, err := s.getSequencesToSendX1(ctx)
+	if err != nil || len(sequences) == 0 {
+		if err != nil {
+			log.Errorf("error getting sequences: %v", err)
+		} else {
+			log.Info("waiting for sequences to be worth sending to L1")
+		}
+		waitTick(ctx, ticker)
+		return
+	}
+
+	lastVirtualBatchNum, err := s.state.GetLastVirtualBatchNum(ctx, nil)
+	if err != nil {
+		log.Errorf("failed to get last virtual batch num, err: %v", err)
+		waitTick(ctx, ticker)
+		return
+	}
+
+	// Send sequences to L1
+	sequenceCount := len(sequences)
+	log.Infof(
+		"sending sequences to L1. From batch %d to batch %d",
+		lastVirtualBatchNum+1, lastVirtualBatchNum+uint64(sequenceCount),
+	)
+	metrics.SequencesSentToL1(float64(sequenceCount))
+
+	// add sequence to be monitored
+	signaturesAndAddrs, err := s.getSignaturesAndAddrsFromDataCommittee(ctx, sequences)
+	if err != nil {
+		log.Error("error getting signatures and addresses from the data committee: ", err)
+		waitTick(ctx, ticker)
+		return
+	}
+	if !s.isValidium() {
+		signaturesAndAddrs = nil
+	}
+	to, data, err := s.etherman.BuildSequenceBatchesTxDataX1(s.cfg.SenderAddress, sequences, s.cfg.L2Coinbase, signaturesAndAddrs)
+	if err != nil {
+		log.Error("error estimating new sequenceBatches to add to eth tx manager: ", err)
+		waitTick(ctx, ticker)
+		return
+	}
+	firstSequence := sequences[0]
+	lastSequence := sequences[len(sequences)-1]
+	monitoredTxID := fmt.Sprintf(monitoredIDFormat, firstSequence.BatchNumber, lastSequence.BatchNumber)
+	err = s.ethTxManager.Add(ctx, ethTxManagerOwner, monitoredTxID, s.cfg.SenderAddress, to, nil, data, s.cfg.GasOffset, nil)
+	if err != nil {
+		mTxLogger := ethtxmanager.CreateLogger(ethTxManagerOwner, monitoredTxID, s.cfg.SenderAddress, to)
+		mTxLogger.Errorf("error to add sequences tx to eth tx manager: ", err)
+		waitTick(ctx, ticker)
+		return
+	}
+}
+
+// getSequencesToSend generates an array of sequences to be send to L1.
+// If the array is empty, it doesn't necessarily mean that there are no sequences to be sent,
+// it could be that it's not worth it to do so yet.
+func (s *SequenceSender) getSequencesToSendX1(ctx context.Context) ([]types.Sequence, error) {
+	lastVirtualBatchNum, err := s.state.GetLastVirtualBatchNum(ctx, nil)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get last virtual batch num, err: %w", err)
+	}
+
+	currentBatchNumToSequence := lastVirtualBatchNum + 1
+	sequences := []types.Sequence{}
+	// var estimatedGas uint64
+	var tx *ethTypes.Transaction
+
+	// Add sequences until too big for a single L1 tx or last batch is reached
+	for {
+		//Check if the next batch belongs to a new forkid, in this case we need to stop sequencing as we need to
+		//wait the upgrade of forkid is completed and s.cfg.NumBatchForkIdUpgrade is disabled (=0) again
+		if (s.cfg.ForkUpgradeBatchNumber != 0) && (currentBatchNumToSequence == (s.cfg.ForkUpgradeBatchNumber + 1)) {
+			return nil, fmt.Errorf("aborting sequencing process as we reached the batch %d where a new forkid is applied (upgrade)", s.cfg.ForkUpgradeBatchNumber+1)
+		}
+
+		// Check if batch is closed
+		isClosed, err := s.state.IsBatchClosed(ctx, currentBatchNumToSequence, nil)
+		if err != nil {
+			return nil, err
+		}
+		if !isClosed {
+			// Reached current (WIP) batch
+			break
+		}
+		// Add new sequence
+		batch, err := s.state.GetBatchByNumber(ctx, currentBatchNumToSequence, nil)
+		if err != nil {
+			return nil, err
+		}
+
+		seq := types.Sequence{
+			GlobalExitRoot: batch.GlobalExitRoot,
+			Timestamp:      batch.Timestamp.Unix(),
+			BatchL2Data:    batch.BatchL2Data,
+			BatchNumber:    batch.BatchNumber,
+		}
+
+		if batch.ForcedBatchNum != nil {
+			forcedBatch, err := s.state.GetForcedBatch(ctx, *batch.ForcedBatchNum, nil)
+			if err != nil {
+				return nil, err
+			}
+			seq.ForcedBatchTimestamp = forcedBatch.ForcedAt.Unix()
+		}
+
+		sequences = append(sequences, seq)
+		if s.isValidium() {
+			if len(sequences) == int(s.cfg.MaxBatchesForL1) {
+				log.Infof(
+					"sequence should be sent to L1, because MaxBatchesForL1 (%d) has been reached",
+					s.cfg.MaxBatchesForL1,
+				)
+				return sequences, nil
+			}
+		} else {
+			// Check if can be send
+			tx, err = s.etherman.EstimateGasSequenceBatchesX1(s.cfg.SenderAddress, sequences, s.cfg.L2Coinbase, nil)
+			if err == nil && tx.Size() > s.cfg.MaxTxSizeForL1 {
+				metrics.SequencesOvesizedDataError()
+				log.Infof("oversized Data on TX oldHash %s (txSize %d > %d)", tx.Hash(), tx.Size(), s.cfg.MaxTxSizeForL1)
+				err = ErrOversizedData
+			}
+			if err != nil {
+				log.Infof("Handling estimage gas send sequence error: %v", err)
+				sequences, err = s.handleEstimateGasSendSequenceErr(ctx, sequences, currentBatchNumToSequence, err)
+				if sequences != nil {
+					// Handling the error gracefully, re-processing the sequence as a sanity check
+					_, err = s.etherman.EstimateGasSequenceBatchesX1(s.cfg.SenderAddress, sequences, s.cfg.L2Coinbase, nil)
+					return sequences, err
+				}
+				return sequences, err
+			}
+		}
+
+		//Check if the current batch is the last before a change to a new forkid, in this case we need to close and send the sequence to L1
+		if (s.cfg.ForkUpgradeBatchNumber != 0) && (currentBatchNumToSequence == (s.cfg.ForkUpgradeBatchNumber)) {
+			log.Info("sequence should be sent to L1, as we have reached the batch %d from which a new forkid is applied (upgrade)", s.cfg.ForkUpgradeBatchNumber)
+			return sequences, nil
+		}
+
+		// Increase batch num for next iteration
+		currentBatchNumToSequence++
+	}
+
+	// Reached latest batch. Decide if it's worth to send the sequence, or wait for new batches
+	if len(sequences) == 0 {
+		log.Info("no batches to be sequenced")
+		return nil, nil
+	}
+
+	lastBatchVirtualizationTime, err := s.state.GetTimeForLatestBatchVirtualization(ctx, nil)
+	if err != nil && !errors.Is(err, state.ErrNotFound) {
+		log.Warnf("failed to get last l1 interaction time, err: %v. Sending sequences as a conservative approach", err)
+		return sequences, nil
+	}
+	if lastBatchVirtualizationTime.Before(time.Now().Add(-s.cfg.LastBatchVirtualizationTimeMaxWaitPeriod.Duration)) {
+		// TODO: implement check profitability
+		// if s.checker.IsSendSequencesProfitable(new(big.Int).SetUint64(estimatedGas), sequences) {
+		log.Info("sequence should be sent to L1, because too long since didn't send anything to L1")
+		return sequences, nil
+		//}
+	}
+
+	log.Info("not enough time has passed since last batch was virtualized, and the sequence could be bigger")
+	return nil, nil
+}
+
+func (s *SequenceSender) isValidium() bool {
+	if !s.cfg.UseValidium {
+		return false
+	}
+
+	committee, err := s.etherman.GetCurrentDataCommittee()
+	if err != nil {
+		return false
+	}
+
+	if len(committee.Members) <= 0 {
+		return false
+	}
+	return true
+}
