+++ state/types_x1.go
@@ -1,0 +1,6 @@
+package state
+
+// IsFlatCallTracer returns true when should use flatCallTracer
+func (t *TraceConfig) IsFlatCallTracer() bool {
+	return t.Tracer != nil && *t.Tracer == "flatCallTracer"
+}
