+++ synchronizer/metrics/metrics.go
@@ -37,15 +37,6 @@ const (
 
 	// ProcessTrustedBatchTimeName is the name of the label to process trusted batch.
 	ProcessTrustedBatchTimeName = Prefix + "process_trusted_batch_time"
-
-	// VirtualBatchNumName is the name of the metric virtual batch number
-	VirtualBatchNumName = Prefix + "virtual_batch_num"
-
-	// VerifiedBatchNumName is the name of the metric verified batch number
-	VerifiedBatchNumName = Prefix + "verified_batch_num"
-
-	// HaltCountName is the name of the metric that counts synchronizer halt count
-	HaltCountName = Prefix + "halt_count"
 )
 
 // Register the metrics for the synchronizer package.
@@ -88,36 +79,10 @@ func Register() {
 			Help: "[SYNCHRONIZER] process trusted batch time",
 		},
 	}
-	gauge := []prometheus.GaugeOpts{
-		{
-			Name: VirtualBatchNumName,
-			Help: "[SYNCHRONIZER] virtual batch num",
-		},
-		{
-			Name: VerifiedBatchNumName,
-			Help: "[SYNCHRONIZER] verified batch num",
-		},
-	}
-	counters := []prometheus.CounterOpts{
-		{
-			Name: HaltCountName,
-			Help: "[SYNCHRONIZER] total count of halt",
-		},
-	}
 
 	metrics.RegisterGauges(gauge...)
 	metrics.RegisterHistograms(histograms...)
 	metrics.RegisterCounters(counters...)
-}
-
-// VirtualBatchNum set the gauge to the given virtual batch num
-func VirtualBatchNum(batchNum uint64) {
-	metrics.GaugeSet(VirtualBatchNumName, float64(batchNum))
-}
-
-// VerifiedBatchNum set the gauge to the given verified batch num
-func VerifiedBatchNum(batchNum uint64) {
-	metrics.GaugeSet(VerifiedBatchNumName, float64(batchNum))
 }
 
 // InitializationTime observes the time initializing the synchronizer on the histogram.
@@ -172,9 +137,4 @@ func GetTrustedBatchInfoTime(lastProcessTime time.Duration) {
 func ProcessTrustedBatchTime(lastProcessTime time.Duration) {
 	execTimeInSeconds := float64(lastProcessTime) / float64(time.Second)
 	metrics.HistogramObserve(ProcessTrustedBatchTimeName, execTimeInSeconds)
-}
-
-// HaltCount increases the counter for the synchronizer halt count.
-func HaltCount() {
-	metrics.CounterAdd(HaltCountName, 1)
 }
